feat: 매니저 역할 분리 및 간단한 개선
- 관리자 페이지에서 read만 가능한 매니저 역할군 분리
- 백엔드 api 응답 스타일 변화로 오류 처리 방식 변경
- 누락되었던 rememberme 추가
- 유저의 마지막 접속 시간 표시
- 미디어의 미디어 타입과 조회수 표시

Close #41

+++ src/components/admin/media_manage/MediaDetail.jsx
@@ -1,8 +1,11 @@
 import { useEffect, useState } from "react";
+import { useAdminSession } from "../AdminContext";
 
 function MediaDetail({ mediaId, onBack, onEdit, onManageEpisode }) {
   const [media, setMedia] = useState(null);
   const [showDeleteModal, setShowDeleteModal] = useState(false);
+
+  const { session } = useAdminSession();
 
   useEffect(() => {
     const fetchMedia = async () => {
@@ -10,8 +13,8 @@ function MediaDetail({ mediaId, onBack, onEdit, onManageEpisode }) {
         const res = await fetch(`/api/media/${mediaId}`, {
           credentials: "include",
         });
-        const result = await res.json();
-        const data = result.data;
+        const body = await res.json();
+        const data = body.data;
 
         const categoryNames = data.categories.map((category) => category.name);
         setMedia({ ...data, categoryNames });
@@ -24,6 +27,12 @@ function MediaDetail({ mediaId, onBack, onEdit, onManageEpisode }) {
   }, [mediaId]);
 
   const handleDelete = async () => {
+    if (!session.admin) {
+      alert("권한이 없습니다.");
+      setShowDeleteModal(false);
+      return;
+    }
+
     try {
       const res = await fetch(`/api/media/${mediaId}`, {
         method: "DELETE",
@@ -69,6 +78,7 @@ function MediaDetail({ mediaId, onBack, onEdit, onManageEpisode }) {
         {media && media.categoryNames && (
           <p><strong>카테고리:</strong> {media.categoryNames.join(", ")}</p>
         )}
+        <p><strong>조회수:</strong> {media.totalViewCount }</p>
 
         <button
           style={{
